next intermediate proof of repr->excisive
The tricky bit that I was sort of smelling but couldn't quite pin down before was the difference between p' and p'' - in the k=1 case they are equal, so I didn't see that before.

This still isn't a complete proof, but it feels like it will turn into one once I figure out the right pieces to add to it, while before there was that lurking error stopping progress.