stronger def of exc, proof of Pi_k(G|p) sub Pi_k(G)|_p
Turns out the really difficult proof I was working on was actually a proof of the *opposite* direction of inclusion than I thought it was, and the direction I thought I was working on had a pretty straightforward proof, once we had the right definition of a restriction of a set partition. So now we can give a stronger definition of excisiveness, which might be useful.