finish representable->excisive proof
After an endless amount of wrangling back and forth, I think this is quite possibly the right proof.

There is still something funny going on with the definition of a morphism in Pset being too weak - so I may still have to go back here to edit this. But I suspect the place it will really come up is in the next section, for the reverse implication.