remove hierarchical section
I was sceptical about this already last time around. Now, I think I see much more clearly why it doesn't quite work - the weighting operation works essentially on the edge level in Phi_Rset, not on the hyperedges. So if two vertices share many hyperedges they might be kept by the scissors thing even if none of the hyperedges are kept - so this seems difficult to phrase in the current formalism.

Since it doesn't seem like the most interesting part of the paper anyway, it seems safe to just cut this section.