fix problem with loops at every vertex
I think this is the cleanest place to deal with the issue - just define Pi_k to remember isolated vertices, instead of forcing loops to make there be no isolated vertices. This also lets us drop the requirement that partitions be covers - we never actually used it, it was just a consequence of the loops everywhere definition that we could require it.